requirements.txt
alembic==1.14.0
aniso8601==9.0.1
attrs==24.2.0
autopep8==2.3.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.6
MarkupSafe==3.0.2
mysqlclient==2.2.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uuid7==0.1.0
Werkzeug==3.1.2
